Checksum check, header cleanup and input documentation
Dude I got distracted by the bridge object again so this might be a little bit incorrect or unfinished, sorry bout that.
Turns out the bridge is still buggy in S3K but you can only get crashes from it if the object slot it takes up at the end gets occupied by another objects because a second bridge despawned it.